requirements.txt
Django==2.1.7
google-cloud-datastore==1.7.3
